// document-models/scope-of-work/v1/gen/schema/types.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Address: { input: `${string}:0x${string}`; output: `${string}:0x${string}` };
  Amount: {
    input: { unit?: string; value?: number };
    output: { unit?: string; value?: number };
  };
  Amount_Crypto: {
    input: { unit: string; value: string };
    output: { unit: string; value: string };
  };
  Amount_Currency: {
    input: { unit: string; value: string };
    output: { unit: string; value: string };
  };
  Amount_Fiat: {
    input: { unit: string; value: number };
    output: { unit: string; value: number };
  };
  Amount_Money: { input: number; output: number };
  Amount_Percentage: { input: number; output: number };
  Amount_Tokens: { input: number; output: number };
  AttachmentRef: {
    input: `attachment://v${number}:${string}`;
    output: `attachment://v${number}:${string}`;
  };
  Currency: { input: string; output: string };
  Date: { input: string; output: string };
  DateTime: { input: string; output: string };
  EmailAddress: { input: string; output: string };
  EthereumAddress: { input: string; output: string };
  OID: { input: string; output: string };
  OLabel: { input: string; output: string };
  PHID: { input: string; output: string };
  URL: { input: string; output: string };
  Unknown: { input: unknown; output: unknown };
  Upload: { input: File; output: File };
};
// ...
export type Agent = {
  description: Maybe<Scalars["String"]["output"]>;
  icon: Maybe<Scalars["URL"]["output"]>;
  id: Scalars["PHID"]["output"];
  name: Scalars["String"]["output"];
};

export type BudgetAnchorProject = {
  margin: Scalars["Float"]["output"];
  marginPinned: Maybe<Scalars["Boolean"]["output"]>;
  project: Maybe<Scalars["OID"]["output"]>;
  quantity: Scalars["Float"]["output"];
  unit: Maybe<Unit>;
  unitCost: Scalars["Float"]["output"];
};

export type BudgetExpenditure = {
  actuals: Scalars["Float"]["output"];
  cap: Scalars["Float"]["output"];
  percentage: Scalars["Float"]["output"];
};

export type BudgetType = "CAPEX" | "CONTINGENCY" | "OPEX" | "OVERHEAD";

export type Deliverable = {
  budgetAnchor: Maybe<BudgetAnchorProject>;
  code: Scalars["String"]["output"];
  description: Scalars["String"]["output"];
  goalRef: Maybe<Scalars["OID"]["output"]>;
  icon: Maybe<Scalars["String"]["output"]>;
  id: Scalars["OID"]["output"];
  keyResults: Array<KeyResult>;
  owner: Maybe<Scalars["ID"]["output"]>;
  status: DeliverableStatus;
  title: Scalars["String"]["output"];
  workProgress: Maybe<Progress>;
};

export type DeliverableSetStatus =
  | "CANCELED"
  | "DRAFT"
  | "FINISHED"
  | "IN_PROGRESS"
  | "TODO";

export type DeliverableSetStatusInput =
  | "CANCELED"
  | "DRAFT"
  | "FINISHED"
  | "IN_PROGRESS"
  | "TODO";

export type DeliverableStatus =
  | "BLOCKED"
  | "CANCELED"
  | "DELIVERED"
  | "DRAFT"
  | "IN_PROGRESS"
  | "TODO"
  | "WONT_DO";

export type DeliverablesCompleted = {
  completed: Scalars["Int"]["output"];
  total: Scalars["Int"]["output"];
};

export type DeliverablesCompletedInput = {
  completed: Scalars["Int"]["input"];
  total: Scalars["Int"]["input"];
};

export type DeliverablesSet = {
  deliverables: Array<Scalars["OID"]["output"]>;
  deliverablesCompleted: DeliverablesCompleted;
  progress: Progress;
  status: DeliverableSetStatus;
};
// ...
export type KeyResult = {
  id: Scalars["OID"]["output"];
  link: Scalars["String"]["output"];
  title: Scalars["String"]["output"];
};

export type LinkDeliverableGoalInput = {
  deliverableId: Scalars["OID"]["input"];
  goalRef?: InputMaybe<Scalars["OID"]["input"]>;
};

export type LinkProjectWbsInput = {
  projectId: Scalars["OID"]["input"];
  wbsRef?: InputMaybe<Scalars["PHID"]["input"]>;
};

export type Milestone = {
  budget: Maybe<Scalars["Float"]["output"]>;
  coordinators: Array<Scalars["ID"]["output"]>;
  deliveryTarget: Scalars["String"]["output"];
  description: Scalars["String"]["output"];
  id: Scalars["OID"]["output"];
  scope: Maybe<DeliverablesSet>;
  sequenceCode: Scalars["String"]["output"];
  title: Scalars["String"]["output"];
};

export type PmBudgetTypeInput = "CAPEX" | "CONTINGENCY" | "OPEX" | "OVERHEAD";

export type PmCurrency = "DAI" | "EUR" | "USD" | "USDS";

export type PmCurrencyInput = "DAI" | "EUR" | "USD" | "USDS";

export type PmDeliverableStatusInput =
  | "BLOCKED"
  | "CANCELED"
  | "DELIVERED"
  | "DRAFT"
  | "IN_PROGRESS"
  | "TODO"
  | "WONT_DO";

export type Progress = {
  completed: Maybe<Scalars["Int"]["output"]>;
  done: Maybe<Scalars["Boolean"]["output"]>;
  total: Maybe<Scalars["Int"]["output"]>;
  value: Maybe<Scalars["Float"]["output"]>;
};

export type ProgressInput = {
  done?: InputMaybe<Scalars["Boolean"]["input"]>;
  percentage?: InputMaybe<Scalars["Float"]["input"]>;
  storyPoints?: InputMaybe<StoryPointInput>;
};

export type Project = {
  abstract: Maybe<Scalars["String"]["output"]>;
  budget: Maybe<Scalars["Float"]["output"]>;
  budgetType: Maybe<BudgetType>;
  code: Scalars["String"]["output"];
  currency: Maybe<PmCurrency>;
  expenditure: Maybe<BudgetExpenditure>;
  id: Scalars["OID"]["output"];
  imageUrl: Maybe<Scalars["URL"]["output"]>;
  knowledgeRefs: Array<Scalars["PHID"]["output"]>;
  projectOwner: Maybe<Scalars["ID"]["output"]>;
  references: Array<Scalars["URL"]["output"]>;
  scope: Maybe<DeliverablesSet>;
  slug: Scalars["String"]["output"];
  targetBudget: Maybe<Scalars["Float"]["output"]>;
  title: Scalars["String"]["output"];
  wbsRef: Maybe<Scalars["PHID"]["output"]>;
};
// ...
export type Roadmap = {
  description: Scalars["String"]["output"];
  id: Scalars["OID"]["output"];
  milestones: Array<Milestone>;
  slug: Scalars["String"]["output"];
  title: Scalars["String"]["output"];
};

export type ScopeOfWorkState = {
  contributors: Array<Agent>;
  deliverables: Array<Deliverable>;
  description: Scalars["String"]["output"];
  projects: Array<Project>;
  roadmaps: Array<Roadmap>;
  status: ScopeOfWorkStatus;
  title: Scalars["String"]["output"];
};

export type ScopeOfWorkStatus =
  | "APPROVED"
  | "CANCELED"
  | "DELIVERED"
  | "DRAFT"
  | "IN_PROGRESS"
  | "REJECTED"
  | "SUBMITTED";
// ...
export type Unit = "Hours" | "StoryPoints";